Make HOUSETYPE, ANGLETYPE, CMDTYPE, and DropLocation "enum class".

+++ src/Choam.cpp
@@ -48,7 +48,7 @@ int Choam::getPrice(Uint32 itemID) const {
 }
 
 bool Choam::isCheap(Uint32 itemID) const {
-    return (getPrice(itemID) < currentGame->objectData.data[itemID][house->getHouseID()].price * 1.3_fix); // A bit of logic to make starports better
+    return (getPrice(itemID) < currentGame->objectData.data[itemID][static_cast<int>(house->getHouseID())].price * 1.3_fix); // A bit of logic to make starports better
 }
 
 int Choam::getNumAvailable(Uint32 itemID) const {
@@ -92,7 +92,7 @@ void Choam::update() {
 
     if((currentGame->getGameCycleCount() % CHOAM_CHANGE_PRICETIME) == 0) {
         for(BuildItem& buildItem : availableItems) {
-            int price = currentGame->objectData.data[buildItem.itemID][house->getHouseID()].price;
+            int price = currentGame->objectData.data[buildItem.itemID][static_cast<int>(house->getHouseID())].price;
 
             const int min_mod = 2;
             const int max_mod = 8;
