+++ src/structures/GunTurret.cpp
@@ -44,7 +44,7 @@ void GunTurret::init() {
     graphic = pGFXManager->getObjPic(ObjPic_GunTurret,getOwner()->getHouseID());
     numImagesX = 10;
     numImagesY = 1;
-    curAnimFrame = firstAnimFrame = lastAnimFrame = ((10-drawnAngle) % 8) + 2;
+    curAnimFrame = firstAnimFrame = lastAnimFrame = ((10-static_cast<int>(drawnAngle)) % 8) + 2;
 }
 
 GunTurret::~GunTurret() = default;
